# Faster R4 legacy quants